[Core] Split Omni connector model runner mixin (#6903)
