+++ vllm_omni/distributed/omni_connectors/model_runner/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM-Omni project
+"""Internal model-runner components for Omni connector transport."""
